feat(vue): refactor api.ts to auth-api.ts
